+++ previous-versions/banking/src/codatbanking/models/operations/list_bank_transactions.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
+from ..shared import errormessage as shared_errormessage
 from ..shared import transactions as shared_transactions
 from typing import Optional
 
@@ -29,6 +30,8 @@ class ListBankTransactionsResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
+    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     transactions: Optional[shared_transactions.Transactions] = dataclasses.field(default=None)